add cron fixes
