+++ frontend/apps/web/src/components/agents/AgentSidebar.tsx
@@ -3,7 +3,8 @@ import type { Agent, AgentMcpConnection, AgentProfile } from '@omnara/sdk'
 import { Link } from '@tanstack/react-router'
 import { useState } from 'react'
 
-import { CreateCronTriggerDialog, CronTriggersList } from '@/components/agents/CronTriggersSection'
+import { CreateCronTriggerDialog } from '@/components/agents/CronTriggerDialog'
+import { CronTriggersList } from '@/components/agents/CronTriggersSection'
 import { DetailList } from '@/components/data-table/DetailList'
 import { InfoIcon, PlusIcon } from '@/components/icons'
 import { Badge } from '@/components/ui/badge'
